requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
twilio
openai
pymongo
firebase-admin
apscheduler
httpx 
python-multipart